Split settings_fields into its own module with a file for each git filetype
