--- src/cnaas_nms/db/settings_fields.py
@@ -1,758 +1,0 @@
-import datetime  # noqa: F401
-import re
-from enum import Enum, StrEnum, auto
-from functools import cached_property
-from ipaddress import AddressValueError, IPv4Address, IPv4Interface, IPv4Network, IPv6Address, IPv6Network
-from typing import Annotated, Dict, List, Literal, Optional, Self
-
-import jmespath
-from aerleon.lib.policy_builder import TermsList
-from jmespath.exceptions import ParseError
-from netutils.lib_mapper import AERLEON_LIB_MAPPER, NAPALM_LIB_MAPPER
-from pydantic import BaseModel, Field, TypeAdapter, ValidationInfo, field_validator, model_validator
-from pydantic.functional_validators import AfterValidator
-
-from cnaas_nms.db.device import Device
-from cnaas_nms.tools.log import get_logger
-
-# HOSTNAME_REGEX = r'([a-z0-9-]{1,63}\.?)+'
-IPV4_REGEX = r"(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\.){3}" r"(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)"
-# IPv6 regex from https://stackoverflow.com/questions/53497/regular-expression-that-matches-valid-ipv6-addresses
-#  minus IPv4 mapped etc since we probably can't handle them anyway
-IPV6_REGEX = (
-    r"(([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}|"  # 1:2:3:4:5:6:7:8
-    r"([0-9a-fA-F]{1,4}:){1,7}:|"  # 1::                              1:2:3:4:5:6:7::
-    r"([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]{1,4}|"  # 1::8             1:2:3:4:5:6::8  1:2:3:4:5:6::8
-    r"([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4}){1,2}|"  # 1::7:8           1:2:3:4:5::7:8  1:2:3:4:5::8
-    r"([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}|"  # 1::6:7:8         1:2:3:4::6:7:8  1:2:3:4::8
-    r"([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|"  # 1::5:6:7:8       1:2:3::5:6:7:8  1:2:3::8
-    r"([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|"  # 1::4:5:6:7:8     1:2::4:5:6:7:8  1:2::8
-    r"[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|"  # 1::3:4:5:6:7:8   1::3:4:5:6:7:8  1::8
-    r":((:[0-9a-fA-F]{1,4}){1,7}|:))"
-)
-HOSTNAME_REGEX = r"^([a-zA-Z0-9-]{1,63})(\.[a-zA-Z-][a-zA-Z0-9-]{0,62})*$"
-HOST_REGEX = f"^({IPV4_REGEX}|{IPV6_REGEX}|{HOSTNAME_REGEX})$"
-DOMAIN_NAME_REGEX = r"^([a-zA-Z0-9-]{1,63})(\.[a-zA-Z0-9-]{1,63})+$"
-host_schema = Field(..., pattern=HOST_REGEX, max_length=253, description="Hostname, FQDN or IP address")
-hostname_schema = Field(..., pattern=HOSTNAME_REGEX, max_length=253, description="Hostname or FQDN")
-domain_name_schema = Field(default=None, pattern=DOMAIN_NAME_REGEX, max_length=251, description="DNS domain name")
-ipv4_schema = Field(..., pattern=f"^{IPV4_REGEX}$", description="IPv4 address")
-IPV4_IF_REGEX = f"{IPV4_REGEX}" + r"\/[0-9]{1,2}"
-ipv4_if_schema = Field(pattern=f"^{IPV4_IF_REGEX}$", description="IPv4 address in CIDR/prefix notation (0.0.0.0/0)")
-ipv6_schema = Field(..., pattern=f"^{IPV6_REGEX}$", description="IPv6 address")
-IPV6_IF_REGEX = f"{IPV6_REGEX}" + r"\/[0-9]{1,3}"
-ipv6_if_schema = Field(pattern=f"^{IPV6_IF_REGEX}$", description="IPv6 address in CIDR/prefix notation (::/0)")
-ipv6_if_schema_optional = Field(
-    default=None, pattern=f"^{IPV6_IF_REGEX}$", description="IPv6 address in CIDR/prefix notation (::/0)"
-)
-ipv4_or_ipv6_if_schema = Field(pattern=f"({IPV4_IF_REGEX}|{IPV6_IF_REGEX})", description="IPv4 or IPv6 prefix")
-
-# VLAN name is alphanumeric max 32 chars on Cisco
-# should not start with number according to some Juniper doc
-VLAN_NAME_REGEX = r"^[a-zA-Z][a-zA-Z0-9-_]{0,31}$"
-vlan_name_schema = Field(
-    pattern=VLAN_NAME_REGEX, description="Max 32 alphanumeric chars, " + "beginning with a non-numeric character"
-)
-vlan_name_schema_optional = Field(
-    default=None,
-    pattern=VLAN_NAME_REGEX,
-    description="Max 32 alphanumeric chars, " + "beginning with a non-numeric character",
-)
-vlan_id_schema = Field(..., gt=0, lt=4096, description="Numeric 802.1Q VLAN ID, 1-4095")
-vlan_id_schema_optional = Field(default=None, gt=0, lt=4096, description="Numeric 802.1Q VLAN ID, 1-4095")
-vxlan_vni_schema = Field(..., gt=0, lt=16777215, description="VXLAN Network Identifier")
-vrf_id_schema = Field(..., gt=0, lt=65536, description="VRF identifier, integer between 1-65535")
-mtu_schema = Field(default=None, ge=68, le=9214, description="MTU (Maximum transmission unit) value between 68-9214")
-as_num_schema = Field(
-    gt=0, lt=4294967296, description="BGP Autonomous System number, 1-4294967295 (asdot notation not supported)"
-)
-as_num_schema_optional = Field(
-    default=None,
-    gt=0,
-    lt=4294967296,
-    description="BGP Autonomous System number, 1-4294967295 (asdot notation not supported)",
-)
-IFNAME_REGEX = r"([a-zA-Z0-9\/\.:-])+"
-ifname_schema = Field(default=None, pattern=f"^{IFNAME_REGEX}$", description="Interface name")
-IFNAME_RANGE_REGEX = r"([a-zA-Z0-9\/\.:\-\[\]])+"
-ifname_range_schema = Field(pattern=f"^{IFNAME_RANGE_REGEX}$", description="Interface range pattern or interface name")
-IFCLASS_REGEX = r"(custom|downlink|fabric|mirror|port_template_[a-zA-Z0-9_]+)"
-ifclass_schema = Field(pattern=f"^{IFCLASS_REGEX}$", description="Interface class: custom, downlink or uplink")
-ifdescr_schema = Field(default=None, max_length=64, description="Interface description, 0-64 characters")
-tcpudp_port_schema = Field(default=None, ge=0, lt=65536, description="TCP or UDP port number, 0-65535")
-ebgp_multihop_schema = Field(default=None, ge=1, le=255, description="Numeric IP TTL, 1-255")
-maximum_routes_schema = Field(
-    default=None, ge=0, le=4294967294, description="Maximum number of routes to receive from peer"
-)
-accept_or_reject_schema = Field(..., pattern=r"^(accept|reject)$", description="Value has to be 'accept' or 'reject'")
-prefix_size_or_range_schema = Field(pattern=r"^[0-9]{1,3}([-][0-9]{1,3})?$", description="Prefix size or range 0-128")
-
-GROUP_NAME = r"^([a-zA-Z0-9_-]{1,63}\.?)+$"
-group_name = Field(..., pattern=GROUP_NAME, max_length=253)
-group_priority_schema = Field(
-    0, ge=0, le=100, description="Group priority 0-100, default 0, higher value means higher priority"
-)
-ACCESS_LIST_NAME = r"^([a-zA-Z0-9_-]{1,63}\.?)+$"
-access_list_name = Annotated[str, Field(pattern=ACCESS_LIST_NAME, max_length=63)]  # Type
-
-
-class RemovePrivateASEnum(StrEnum):
-    ALL = auto()
-    REPLACE = auto()
-
-
-class VlanOptionEnum(StrEnum):
-    NONE = auto()
-    TAGGED = auto()
-    UNTAGGED = auto()
-
-
-def validate_ipv4_if(ipv4if: str):
-    try:
-        assert "/" in ipv4if, "Not a CIDR notation/no netmask"
-        addr = IPv4Interface(ipv4if)
-        assert 8 <= addr.network.prefixlen <= 32, "Invalid prefix size"
-        assert not addr.is_multicast, "Multicast address is invalid"
-        if addr.network.prefixlen <= 30:
-            assert str(addr.ip) != str(addr.network.network_address), "Invalid interface address"
-            assert str(addr.ip) != str(addr.network.broadcast_address), "Invalid interface address"
-    except AddressValueError as e:
-        raise ValueError("Invalid IPv4 interface: {}".format(e))
-    except AssertionError as e:
-        raise ValueError("Invalid IPv4 interface: {}".format(e))
-    return ipv4if
-
-
-# Note: If specifying a list of a BaseModel derived class anywhere else except
-# f_root you will get validation errors, and the errors does not refer to
-# settings from the top level dict but instead use the first level after f_root
-# as their top level. This also makes error messages quite useless.
-# I don't have a solution for this at the moment, so just put list of classes
-# directly under f_root.
-
-
-class f_ntp_server(BaseModel):
-    host: str = host_schema
-
-
-class f_radius_server(BaseModel):
-    host: str = host_schema
-    port: Optional[int] = tcpudp_port_schema
-
-
-class f_syslog_server(BaseModel):
-    host: str = host_schema
-    port: Optional[int] = tcpudp_port_schema
-
-
-class f_flow_collector(BaseModel):
-    host: str = host_schema
-    port: Optional[int] = tcpudp_port_schema
-
-
-class f_snmp_server(BaseModel):
-    host: str = host_schema
-
-
-class f_dns_server(BaseModel):
-    host: str = host_schema
-
-
-class f_dhcp_relay(BaseModel):
-    host: str = host_schema
-
-
-class f_evpn_peer(BaseModel):
-    hostname: str = hostname_schema
-
-
-def vlan_range_check(v: str) -> str:
-    if "-" in v:
-        start, end = v.split("-")
-        assert int(start) < int(end), "Start of range must be less than end of range"
-        assert int(start) >= 1 and int(end) <= 4095, "VLAN IDs in range must be between 1-4095"
-    else:
-        assert 1 <= int(v) <= 4095, "VLAN IDs in range must be between 1-4095"
-    return v
-
-
-def vni_range_required_check(v: str) -> str:
-    if "-" in v:
-        start, end = v.split("-")
-        assert int(start) < int(end), "Start of range must be less than end of range"
-        assert int(start) >= 1 and int(end) <= 16777215, "VNI IDs in range must be between 1-16777215"
-    else:
-        raise ValueError("Range must be specified, ex '10000-99999'")
-    return v
-
-
-class f_interface(BaseModel):
-    name: str = ifname_range_schema
-    ifclass: str = ifclass_schema
-    redundant_link: bool = True
-    config: Optional[str] = None
-    description: Optional[str] = ifdescr_schema
-    enabled: Optional[bool] = None
-    untagged_vlan: Optional[int] = vlan_id_schema_optional
-    # tagged vlan list can be list of vlans IDs or ranges of VLAN IDs ("1-10")
-    tagged_vlan_list: Optional[
-        List[Annotated[int, Field(ge=1, le=4095)] | Annotated[str, AfterValidator(vlan_range_check)]]
-    ] = None
-    aggregate_id: Optional[int] = None
-    tags: Optional[List[str]] = None
-    vrf: Optional[str] = vlan_name_schema_optional
-    ipv4_address: Optional[str] = None
-    ipv6_address: Optional[str] = ipv6_if_schema_optional
-    mtu: Optional[int] = mtu_schema
-    acl_ipv4_in: Optional[str] = None
-    acl_ipv4_out: Optional[str] = None
-    acl_ipv6_in: Optional[str] = None
-    acl_ipv6_out: Optional[str] = None
-    metric: Optional[int] = None
-    cli_append_str: str = ""
-
-    @field_validator("ipv4_address")
-    @classmethod
-    def vrf_required_if_ipv4_address_set(cls, v: str, info: ValidationInfo):
-        if v:
-            validate_ipv4_if(v)
-            if "vrf" not in info.data or not info.data["vrf"]:
-                raise ValueError("VRF is required when specifying ipv4_gw")
-        return v
-
-
-class f_vrf(BaseModel):
-    name: Optional[str] = None
-    vrf_id: int = vrf_id_schema
-    import_route_targets: List[str] = []
-    export_route_targets: List[str] = []
-    import_policy: Optional[str] = None
-    export_policy: Optional[str] = None
-    groups: List[str] = []
-
-
-class f_ipv4_static_route(BaseModel):
-    destination: str = ipv4_if_schema
-    nexthop: str = ipv4_schema
-    interface: Optional[str] = ifname_schema
-    name: str = "undefined"
-    cli_append_str: str = ""
-
-
-class f_ipv6_static_route(BaseModel):
-    destination: str = ipv6_if_schema
-    nexthop: str = ipv6_schema
-    interface: Optional[str] = ifname_schema
-    name: str = "undefined"
-    cli_append_str: str = ""
-
-
-class f_extroute_static_vrf(BaseModel):
-    name: str
-    ipv4: Optional[List[f_ipv4_static_route]] = None
-    ipv6: Optional[List[f_ipv6_static_route]] = None
-
-
-class f_extroute_static(BaseModel):
-    vrfs: List[f_extroute_static_vrf]
-
-
-class f_extroute_ospfv3_vrf(BaseModel):
-    name: str
-    ipv4_redist_routefilter: Optional[str] = None
-    ipv6_redist_routefilter: Optional[str] = None
-    cli_append_str: str = ""
-
-
-class f_extroute_ospfv3(BaseModel):
-    vrfs: List[f_extroute_ospfv3_vrf]
-
-
-class f_extroute_bgp_neighbor_v4(BaseModel):
-    peer_ipv4: str = ipv4_schema
-    peer_as: int = as_num_schema
-    route_map_in: str = vlan_name_schema
-    route_map_out: str = vlan_name_schema
-    description: str = "undefined"
-    bfd: Optional[bool] = None
-    graceful_restart: Optional[bool] = None
-    next_hop_self: Optional[bool] = None
-    update_source: Optional[str] = ifname_schema
-    ebgp_multihop: Optional[int] = ebgp_multihop_schema
-    maximum_routes: Optional[int] = maximum_routes_schema
-    auth_type: Optional[str] = None
-    auth_string: Optional[str] = None
-    remove_private_as: Optional[RemovePrivateASEnum] = None
-    cli_append_str: str = ""
-
-
-class f_extroute_bgp_neighbor_v6(BaseModel):
-    peer_ipv6: str = ipv6_schema
-    peer_as: int = as_num_schema
-    route_map_in: str = vlan_name_schema
-    route_map_out: str = vlan_name_schema
-    description: str = "undefined"
-    bfd: Optional[bool] = None
-    graceful_restart: Optional[bool] = None
-    next_hop_self: Optional[bool] = None
-    update_source: Optional[str] = ifname_schema
-    ebgp_multihop: Optional[int] = ebgp_multihop_schema
-    maximum_routes: Optional[int] = maximum_routes_schema
-    auth_type: Optional[str] = None
-    auth_string: Optional[str] = None
-    remove_private_as: Optional[RemovePrivateASEnum] = None
-    cli_append_str: str = ""
-
-
-class f_extroute_bgp_vrf(BaseModel):
-    name: str
-    local_as: int = as_num_schema
-    neighbor_v4: List[f_extroute_bgp_neighbor_v4] = []
-    neighbor_v6: List[f_extroute_bgp_neighbor_v6] = []
-    cli_append_str: str = ""
-
-
-class f_extroute_bgp(BaseModel):
-    vrfs: List[f_extroute_bgp_vrf] = []
-
-
-class f_internal_vlans(BaseModel):
-    vlan_id_low: int = vlan_id_schema
-    vlan_id_high: int = vlan_id_schema
-    allocation_order: str = "ascending"
-
-    @field_validator("vlan_id_high")
-    @classmethod
-    def vlan_id_high_greater_than_low(cls, v: int, info: ValidationInfo):
-        if v:
-            if info.data["vlan_id_low"] >= v:
-                raise ValueError("vlan_id_high must be greater than vlan_id_low")
-        return v
-
-
-class f_vxlan(BaseModel):
-    description: Optional[str] = None
-    vni: int = vxlan_vni_schema
-    vrf: Optional[str] = vlan_name_schema_optional
-    vlan_id: int = vlan_id_schema
-    vlan_name: str = vlan_name_schema
-    ipv4_gw: Optional[str] = None
-    ipv4_secondaries: Optional[List[Annotated[str, AfterValidator(validate_ipv4_if)]]] = None
-    ipv6_gw: Optional[str] = ipv6_if_schema_optional
-    dhcp_relays: Optional[List[f_dhcp_relay]] = None
-    mtu: Optional[int] = mtu_schema
-    vxlan_host_route: bool = True
-    acl_ipv4_in: Optional[str] = None
-    acl_ipv4_out: Optional[str] = None
-    acl_ipv6_in: Optional[str] = None
-    acl_ipv6_out: Optional[str] = None
-    cli_append_str: str = ""
-    groups: List[str] = []
-    devices: List[str] = []
-    tags: List[str] = []
-
-    @field_validator("ipv4_gw")
-    @classmethod
-    def vrf_required_if_ipv4_gw_set(cls, v: str, info: ValidationInfo):
-        if v:
-            validate_ipv4_if(v)
-            if "vrf" not in info.data or not info.data["vrf"]:
-                raise ValueError("VRF is required when specifying ipv4_gw")
-        return v
-
-    @field_validator("ipv6_gw")
-    @classmethod
-    def vrf_required_if_ipv6_gw_set(cls, v: str, info: ValidationInfo):
-        if v:
-            if "vrf" not in info.data or not info.data["vrf"]:
-                raise ValueError("VRF is required when specifying ipv6_gw")
-        return v
-
-
-class f_underlay(BaseModel):
-    infra_lo_net: str = ipv4_if_schema
-    infra_link_net: str = ipv4_if_schema
-    mgmt_lo_net: str = ipv4_if_schema
-    bgp_asn: Optional[int] = as_num_schema_optional
-
-
-class f_user(BaseModel):
-    username: str
-    ssh_key: Optional[str] = None
-    uid: Optional[int] = None
-    password_hash_arista: Optional[str] = None
-    password_hash_cisco: Optional[str] = None
-    password_hash_juniper: Optional[str] = None
-    permission_arista: Optional[str] = None
-    permission_cisco: Optional[str] = None
-    permission_juniper: Optional[str] = None
-    groups: List[str] = []
-
-
-class f_prefixset_item(BaseModel):
-    prefix: str = ipv4_or_ipv6_if_schema
-    masklength_range: Optional[Annotated[int, Field(ge=0, le=128)] | Annotated[str, prefix_size_or_range_schema]] = None
-
-
-class f_prefixset(BaseModel):
-    mode: str = "ipv4"
-    prefixes: List[f_prefixset_item]
-
-
-class f_rpolicy_condition(BaseModel):
-    match_type: str
-    match_target: str
-
-
-class f_rpolicy_statement(BaseModel):
-    action: str = accept_or_reject_schema
-    conditions: List[f_rpolicy_condition]
-
-
-class f_routingpolicy(BaseModel):
-    statements: List[f_rpolicy_statement]
-
-
-class f_interface_tag(BaseModel):
-    description: str = ""
-    groups: Optional[List[str]] = None
-
-
-class f_port_template(BaseModel):
-    description: str = ""
-    vlan_config: VlanOptionEnum = VlanOptionEnum.TAGGED
-    groups: Optional[List[str]] = None
-
-
-class f_network_definition(BaseModel):
-    address: IPv4Address | IPv6Address | IPv4Network | IPv6Network
-    comment: str = ""
-
-    # Convert address to string.
-    @field_validator("address", mode="after")
-    @classmethod
-    def validate_address(cls, v):
-        return str(v)
-
-
-class f_network_definition_include(BaseModel):
-    name: str
-
-
-class f_network_definition_reference(BaseModel):
-    path: str
-    strip_cidr: Optional[bool] = False
-
-    @field_validator("path", mode="after")
-    @classmethod
-    def validate_jmespath(cls, v: str) -> str:
-        try:
-            jmespath.compile(v)
-        except ParseError as e:
-            raise ValueError(str(e))
-        return v
-
-
-class f_service_definition(BaseModel):
-    port: int | str
-    protocol: str
-
-
-class f_service_definition_include(BaseModel):
-    name: str
-
-
-TermsListAdapter: TypeAdapter = TypeAdapter(TermsList)
-
-TermsListAdapter.rebuild()
-
-
-class f_access_list(BaseModel):
-    comment: str = ""
-    inet_families: List[Literal["ipv4", "ipv6"]] = ["ipv4"]
-    header_map: Dict[str, str] = {}
-    # Example header_map
-    # {"ios": "{ACL_NAME} {INET_FAMILY} noverbose",
-    # "eos": "ACL_NAME extended noverbose"}
-
-    # Uses Aerleon TypedDict
-    terms: TermsList
-
-    @field_validator("inet_families", mode="after")
-    @classmethod
-    def unique_sorted_inet_families(cls, v: List[Literal["ipv4", "ipv6"]]) -> List[Literal["ipv4", "ipv6"]]:
-        """Make sure inet_families are unique and sorted"""
-        return sorted(set(v))
-
-    @field_validator("header_map", mode="after")
-    @classmethod
-    def validate_header_map(cls, v: Dict[str, str]) -> Dict[str, str]:
-        """Make sure header_map only contains valid options"""
-        valid_platforms = list(NAPALM_LIB_MAPPER.keys()) + list(AERLEON_LIB_MAPPER.keys())
-        for k in v.keys():
-            if k not in valid_platforms:
-                raise ValueError(f"{k} must be a valid napalm or aerleon platform")
-
-        return v
-
-    @field_validator("terms", mode="after")
-    def validate_term_names(cls, v):
-        """
-        All terms must have valid names
-        Term name uniqueness is handled in f_root
-        """
-        for term in v:
-            if "include" in term:
-                # This is a PolicyInclude and is validated later in f_root.
-                continue
-
-            term_name = term.get("name")
-
-            if not term_name:
-                raise ValueError("Terms must have a name")
-
-            # Invalid characters is a ValueError
-            if not re.match(r"^[\w-]+$", term_name):
-                raise ValueError(f"Invalid term name: {term_name}")
-
-        return v
-
-    @field_validator("terms", mode="after")
-    @classmethod
-    def validate_terms(cls, terms):
-        if not terms:
-            raise ValueError("Terms must be defined and cannot be empty")
-
-        # Validate all terms regarding to the TypedDict
-        TermsListAdapter.validate_python(terms)
-
-        return terms
-
-
-f_access_list.model_rebuild()
-
-
-class f_group_device_filter(BaseModel):
-    hostname: Optional[str] = None
-    device_type: Optional[str] = None
-    model: Optional[str] = None
-    os_version: Optional[str] = None
-    platform: Optional[str] = None
-
-    @field_validator("hostname", "device_type", "model", "os_version", "platform")
-    @classmethod
-    def validate_regex(cls, v):
-        """Validate that the value is a valid regex pattern."""
-        try:
-            # Try compiling regex
-            re.compile(v)
-        except re.error as exc:
-            raise ValueError(f"Invalid regex pattern: {v}") from exc
-        return v
-
-    @cached_property
-    def compiled_patterns(self) -> Dict[str, re.Pattern]:
-        """
-        Is a cached property to avoid re-compiling regex patterns
-        """
-        fields = set(self.__annotations__.keys())
-        compiled_patterns = {}
-        for field in fields:
-            pattern = getattr(self, field, None)
-            if pattern:
-                compiled_patterns.update({field: re.compile(pattern)})
-        return compiled_patterns
-
-    def matches(self, device: Device) -> bool:
-        """
-        A function that matches a device based on the regex patterns.
-        """
-        compiled_patterns = self.compiled_patterns
-        # No patterns defined, match nothing
-        if not compiled_patterns:
-            return False
-
-        for field, pattern in compiled_patterns.items():
-            value = getattr(device, field, None)
-            if value is None:
-                return False  # field missing → no match
-            if isinstance(value, Enum):
-                match_value = value.name
-            else:
-                match_value = value
-            if not pattern.match(str(match_value)):  # convert to str to be safe
-                return False  # pattern did not match
-        return True  # all matched
-
-
-class f_group(BaseModel):
-    name: str = group_name
-    device_filter: Optional[f_group_device_filter] = None
-    devices: Optional[List[str]] = None
-    group_priority: int = group_priority_schema
-    templates_branch: Optional[str] = None
-
-    def __init__(self, **data):
-        logger = get_logger()
-        if "group" in data:
-            logger.warning(
-                "Old group config style is deprecated and will be removed in a future version.",
-            )
-            legacy_data = data.pop("group")
-            # Convert legacy group data to new format
-            data["name"] = legacy_data.pop("name")
-            regex = legacy_data.pop("regex")
-            if regex:
-                data["device_filter"] = {"hostname": regex}
-            data["group_priority"] = legacy_data.pop("group_priority", 0)
-            data["templates_branch"] = legacy_data.pop("templates_branch", None)
-        super().__init__(**data)
-
-    @field_validator("group_priority")
-    @classmethod
-    def reserved_priority(cls, v: int, info: ValidationInfo):
-        if v and v == 1 and info.data["name"] != "DEFAULT":
-            raise ValueError("group_priority 1 is reserved for built-in group DEFAULT")
-        return v
-
-    @field_validator("templates_branch")
-    @classmethod
-    def templates_branch_primary_group_only(cls, v: str, info: ValidationInfo):
-        if v and info.data["group_priority"] <= 1:
-            raise ValueError("templates_branch can only be specified on primary groups")
-        return v
-
-    @model_validator(mode="after")
-    def cannot_use_device_filter_with_devices(self: Self) -> Self:
-        device_filter = self.device_filter
-        devices = self.devices
-        if device_filter and devices:
-            raise ValueError("cannot use device_filter together with devices")
-
-        return self
-
-    def matches(self, device: Device) -> bool:
-        """A function to check if a device matches the group filters."""
-        if self.device_filter is not None:
-            # Use the device filter matcher to check if the device matches
-            return self.device_filter.matches(device)
-        elif self.devices is not None:
-            # If no device filter is defined, check if the device is in the devices list
-            return device.hostname in self.devices
-        return False
-
-
-def validate_groups(groups: List[f_group]):
-    """
-    Validate that the provided list of groups have unique names and group priorities.
-    """
-
-    # Validate uniqueness of group names and group priorities
-    unique_fields = ["name", "group_priority"]
-    for unique_field in unique_fields:
-        seen = set()
-        for group in groups:
-            value = getattr(group, unique_field)
-            # Skip validation for group_priority if it's 0
-            if unique_field == "group_priority" and value == 0:
-                continue
-            if value in seen:
-                raise ValueError(
-                    f"Groups must have unique {unique_field} values, "
-                    f"but group {group} has a duplicate {unique_field} value as another group."
-                )
-            seen.add(value)
-
-    return groups
-
-
-class f_groups(BaseModel):
-    groups: Annotated[Optional[List[f_group]], AfterValidator(validate_groups)] = None
-
-
-class f_access_lists(BaseModel):
-    network_definitions: Dict[
-        str, List[f_network_definition | f_network_definition_include | f_network_definition_reference]
-    ] = {}
-    service_definitions: Dict[str, List[f_service_definition | f_service_definition_include]] = {}
-    access_lists: Dict[access_list_name, f_access_list] = {}
-
-    @field_validator("access_lists", mode="after")
-    @classmethod
-    def validate_access_lists_includes(
-        cls, access_lists: Dict[access_list_name, f_access_list]
-    ) -> Dict[access_list_name, f_access_list]:
-        """Raise an error if some term include is not pointing to a valid access_list"""
-        acl_names = access_lists.keys()
-        for access_list in access_lists.values():
-            for term in access_list.terms:
-                include_acl = term.get("include")
-                if include_acl and include_acl not in acl_names:
-                    raise ValueError(f"Included access-list: {include_acl} must be defined.")
-        return access_lists
-
-    @field_validator("access_lists", mode="after")
-    @classmethod
-    def validate_access_lists_included_terms(
-        cls, access_lists: Dict[access_list_name, f_access_list]
-    ) -> Dict[access_list_name, f_access_list]:
-        """Validates an access-list + included access-lists have unique term-names"""
-        for access_list in access_lists.values():
-            all_term_names = [t.get("name") for t in access_list.terms if t.get("name")]
-            for term in access_list.terms:
-                include_acl = term.get("include")
-
-                if include_acl and isinstance(include_acl, str):
-                    all_term_names.extend([t.get("name") for t in access_lists[include_acl].terms if t.get("name")])
-
-            if len(all_term_names) != len(set(all_term_names)):
-                raise ValueError("All term names in an access-list + included access-lists must be unique.")
-        return access_lists
-
-
-class f_base_system(BaseModel):
-    ntp_servers: List[f_ntp_server] = []
-    radius_servers: List[f_radius_server] = []
-    syslog_servers: List[f_syslog_server] = []
-    snmp_servers: List[f_snmp_server] = []
-    dns_servers: List[f_dns_server] = []
-    flow_collectors: List[f_flow_collector] = []
-    dhcp_relays: Optional[List[f_dhcp_relay]] = None
-    internal_vlans: Optional[f_internal_vlans] = None
-    dot1x_fail_vlan: Optional[int] = vlan_id_schema_optional
-    cli_prepend_str: str = ""
-    cli_append_str: str = ""
-    organization_name: str = ""
-    domain_name: Optional[str] = domain_name_schema
-    users: List[f_user] = []
-    dot1x_multi_host: bool = False
-    poe_reboot_maintain: bool = False
-    interface_tag_options: Dict[str, f_interface_tag] = {}
-    port_template_options: Dict[str, f_port_template] = {}
-    vxlan_vni_range: Optional[Annotated[str, AfterValidator(vni_range_required_check)]] = None
-    arista_models_32bit: Optional[List[str]] = None
-    upgrade_post_waittime: Dict[str, int] = {"default": 600}
-    system_access_lists: List[access_list_name] = []
-    # This is defined both in f_base_system and f_routing
-    external_routing_policies: List[str] = []
-
-
-class f_interfaces(BaseModel):
-    interfaces: List[f_interface] = []
-
-
-class f_routing(BaseModel):
-    vrfs: List[f_vrf] = []
-    underlay: Optional[f_underlay] = None
-    evpn_peers: List[f_evpn_peer] = []
-    extroute_static: Optional[f_extroute_static] = None
-    extroute_ospfv3: Optional[f_extroute_ospfv3] = None
-    extroute_bgp: Optional[f_extroute_bgp] = None
-    prefix_sets: Dict[str, f_prefixset] = {}
-    routing_policies: Dict[str, f_routingpolicy] = {}
-    # This is defined both in f_base_system and f_routing
-    external_routing_policies: List[str] = []
-
-
-class f_vxlans(BaseModel):
-    vxlans: Dict[str, f_vxlan] = {}
