+++ src/cnaas_nms/db/settings_fields/__init__.py
@@ -1,0 +1,8 @@
+# Import all here so we can import directly from settings_fields
+from cnaas_nms.db.settings_fields.access_lists import *  # noqa: F403
+from cnaas_nms.db.settings_fields.base_system import *  # noqa: F403
+from cnaas_nms.db.settings_fields.groups import *  # noqa: F403
+from cnaas_nms.db.settings_fields.interfaces import *  # noqa: F403
+from cnaas_nms.db.settings_fields.routing import *  # noqa: F403
+from cnaas_nms.db.settings_fields.shared import *  # noqa: F403
+from cnaas_nms.db.settings_fields.vxlans import *  # noqa: F403
